Include tiny_queue.h in tiny_queue.m
In order to avoid declaring tiny_queue_t and tiny_linked_list_t twice,
this commit removes the struct typedefs from tiny_queue.m and instead
pulls in the header file tiny_queue.h.
